+++ gpt_researcher/master/prompts.py
@@ -220,3 +220,21 @@ def generate_report_introduction(question: str, research_summary: str = "") -> s
         - You must include hyperlinks with markdown syntax ([url website](url)) related to the sentences wherever necessary.
         Assume that the current date is {datetime.now(timezone.utc).strftime('%B %d, %Y')} if required.
     """
+
+
+report_type_mapping = {
+    ReportType.ResearchReport.value: generate_report_prompt,
+    ReportType.ResourceReport.value: generate_resource_report_prompt,
+    ReportType.OutlineReport.value: generate_outline_report_prompt,
+    ReportType.CustomReport.value: generate_custom_report_prompt,
+    ReportType.SubtopicReport.value: generate_subtopic_report_prompt
+}
+
+
+def get_prompt_by_report_type(report_type):
+    default_report_type = ReportType.ResearchReport.value
+    prompt_by_type = report_type_mapping.get(report_type)
+    if not prompt_by_type:
+        prompt_by_type = report_type_mapping.get(default_report_type)
+        print(f"Invalid report type: {report_type}. Using default report type: {default_report_type}")
+    return prompt_by_type
